API changes Permission

+++ backend/server/auth/permission/permission.controller.ts
@@ -5,13 +5,19 @@ import { Constants, PermissionS, Permission as PermissionEntity } from 'fivebyon
 import { AuthUtil } from '../../util/auth.util';
 
 
-const { HTTP_OK, HTTP_BAD_REQUEST } = Constants;
+const { HTTP_OK, HTTP_BAD_REQUEST, HTTP_UNAUTHORIZED } = Constants;
 
 const router = expressRouter();
 
 const listAll = async(request: any, response: any) => {
 
-  const Permission = PermissionModel.createModel();
+  const sessionDetails = AuthUtil.findSessionDetails(request);
+  if (!sessionDetails.company) {
+
+    return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+  }
+  const Permission = PermissionModel.createModel(sessionDetails.company);
   const permissions = await Permission.find();
   return response.status(HTTP_OK).json(permissions);
 
@@ -22,7 +28,13 @@ const getPermission = async(request: any, response: any) => {
 
   try {
 
-    const Permission = PermissionModel.createModel();
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Permission = PermissionModel.createModel(sessionDetails.company);
     const permission: PermissionEntity = await Permission.findById(request.params.id);
     if (!permission) {
 
@@ -43,7 +55,13 @@ const savePermission = async(request: any, response: any) => {
 
   try {
 
-    const Permission = PermissionModel.createModel();
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Permission = PermissionModel.createModel(sessionDetails.company);
     const permission = new Permission(request.body);
     await permission.save();
     return response.status(HTTP_OK).json(permission);
@@ -70,7 +88,13 @@ const updatePermission = async(request: any, response: any) => {
 
     }
 
-    const Permission = PermissionModel.createModel();
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Permission = PermissionModel.createModel(sessionDetails.company);
     await Permission.updateOne({ _id: id }, permissionObj, { runValidators: true });
     return response.status(HTTP_OK).json(permissionObj);
 
@@ -88,7 +112,13 @@ const deletePermission = async(request: any, response: any) => {
 
     const { id } = request.params;
 
-    const Permission = PermissionModel.createModel();
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Permission = PermissionModel.createModel(sessionDetails.company);
     const resp = await Permission.deleteOne({ _id: id });
     if (resp.deletedCount === 0) {
